+++ src/com/sh/domains/ladders/BothTrue.java
@@ -1,0 +1,6 @@
+package com.sh.domains.ladders;
+
+@FunctionalInterface
+public interface BothTrue {
+	boolean isAnd(boolean a, boolean b);
+}
