+++ test/com/sh/domains/ladders/LadderServiceTest.java
@@ -1,5 +1,6 @@
-package com.sh.domains;
+package com.sh.domains.ladders;
 
+import static com.sh.domains.players.PlayersTest.*;
 import static com.sh.views.InputVerification.*;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -13,19 +14,19 @@ import org.junit.jupiter.api.Test;
 
 import com.sh.domains.ladders.Ladder;
 import com.sh.domains.ladders.LadderService;
-import com.sh.domains.players.Players;
 
 class LadderServiceTest {
 	public static final String TEST_SYMBOL_LADDER = "-----";
+	public static final int VERTICAL_LENGTH_OF_LADDER = 5;
 	private Ladder ladder;
 	private LadderService ladderService;
 
 	static Predicate<String> isNotNull = isStringBlank.negate();
 
-/*	@BeforeEach
+	@BeforeEach
 	void beforeEach() {
 		List<String> names = getNames();
-		ladder = new Ladder(names.size(), 5);
+		ladder = new Ladder(names.size(), VERTICAL_LENGTH_OF_LADDER);
 		ladderService = new LadderService();
 	}
 
@@ -33,13 +34,23 @@ class LadderServiceTest {
 	@Test
 	void acceptable_the_result_of_ladder_length_for_print_out() {
 		String regex = "[\\s\\|]";
-		String ladderText = ladderService.resultOfPlay();
+		String ladderText = ladderService.resultOfLadder(ladder);
 
 		String actual = pickLineAndMaxText(1, regex, ladderText);
 
 		assertEquals(TEST_SYMBOL_LADDER, actual);
-	}*/
+	}
 
+	/**
+	 * 	출력문이 세로순으로는(y축) 플레이어들 이름, 사다리, 사디리 결과 순입니다.
+	 * 	- 첫째줄 이용한 이름 테스트
+	 * 	- 둘째줄 이후 이용한 사다리 라인 확인 테스트
+	 * 	- 마지막줄 이용한 사다리 결과 테스트
+	 * @param idx
+	 * @param regex
+	 * @param ladderText
+	 * @return
+	 */
 	public static String pickLineAndMaxText(int idx, String regex, String ladderText) {
 		String line = ladderText.split(System.lineSeparator())[idx];
 		String separated = line.replaceAll(regex, ",");
@@ -54,5 +65,4 @@ class LadderServiceTest {
 			.max((a, b) -> (a.length() - b.length()))
 			.get();
 	}
-
 }
