+++ test/com/sh/domains/ladders/LineTest.java
@@ -1,4 +1,4 @@
-package com.sh.domains;
+package com.sh.domains.ladders;
 
 import static java.util.stream.Collectors.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.Test;
 import com.sh.domains.ladders.Line;
 
 class LineTest {
-
 	private int numberOfPeople = 5;
 	private List<Boolean> filledFalseList;
